requirements.txt
flake8==3.8.2
pyyaml>=5.1
tensorboard
tensorboardX
matplotlib
onnxruntime
